floss: Align the SetVolume signature
BUG: 194853979
Tag: #floss
Test: build, deploy, connect and test by sending DBus message
BYPASS_LONG_LINES_REASON: Bluetooth likes 120 char lines

+++ system/gd/rust/linux/service/src/iface_bluetooth_media.rs
@@ -121,7 +121,7 @@ impl IBluetoothMedia for IBluetoothMediaDBus {
     }
 
     #[dbus_method("SetVolume")]
-    fn set_volume(&mut self, volume: i32) {
+    fn set_volume(&mut self, volume: u8) {
         dbus_generated!()
     }
